backend/requirements.txt
fastapi
uvicorn
pdfplumber
openai
playwright
python-dotenv
python-multipart